Fix clang-reported warnings

+++ signal/src/decoupling.cpp
@@ -193,7 +193,7 @@ void CDecoupling_Filter::Flush_Stats() {
     }
 
 
-    auto flush_marker = [&stats_file](const size_t segment_id, auto& series, auto& stats, const wchar_t* marker_string) {
+    auto flush_marker = [&stats_file](const uint64_t segment_id, auto& series, auto& stats, const wchar_t* marker_string) {
         scgms::TSignal_Stats signal_stats;
 
         const bool result = Calculate_Signal_Stats(series, signal_stats);
